HBX-1664: Move class 'org.hibernate.tool.hbm2x.pojo.EntityPOJOClass' to 'org.hibernate.tool.internal.export.pojo.EntityPOJOClass'
