+++ test/nodb/src/test/java/org/hibernate/tool/hbm2x/Hbm2JavaEjb3Test/TestCase.java
@@ -15,8 +15,8 @@ import org.hibernate.boot.Metadata;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
-import org.hibernate.tool.hbm2x.pojo.EntityPOJOClass;
 import org.hibernate.tool.internal.export.pojo.Cfg2JavaTool;
+import org.hibernate.tool.internal.export.pojo.EntityPOJOClass;
 import org.hibernate.tool.internal.export.pojo.POJOClass;
 import org.hibernate.tool.internal.export.pojo.POJOExporter;
 import org.hibernate.tool.internal.util.AnnotationBuilder;
